fixed few tests
