+++ pom.xml
@@ -100,8 +100,14 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-surefire-plugin</artifactId>
             <version>3.0.0-M9</version> <!-- Use the latest version -->
+<!--            <configuration>-->
+<!--                <argLine>-Dnet.bytebuddy.experimental=true</argLine>-->
+<!--            </configuration>-->
             <configuration>
-                <argLine>-Dnet.bytebuddy.experimental=true</argLine>
+                <!-- Specify the custom java.security file -->
+                <systemPropertyVariables>
+                    <java.security.properties>${project.basedir}/src/test/resources/custom.java.security</java.security.properties>
+                </systemPropertyVariables>
             </configuration>
         </plugin>                
     </plugins>
