+++ random-provider/src/test/resources/app.properties
@@ -1,0 +1,4 @@
+qrypt.api.url=https://qrypt.com/api
+qrypt.token=1234
+qrypt.store.size=4092
+qrypt.store.min_threshold=409
